requirements.txt
Flask==3.1.3
openai==2.21.0
python-dotenv==1.2.1
requests==2.32.3
gunicorn==23.0.0
opencc-python-reimplemented==0.1.7
